requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.1.2
colorama==0.4.4
dnspython==2.2.1
email-validator==1.1.3
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.0
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-WTF==1.0.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mongoengine==0.24.1
Pillow==9.1.0
pycparser==2.21
pymongo==4.1.1
six==1.16.0
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0